refactor(browser): split copilot custody owners
