+++ .oxfmtrc.jsonc
@@ -32,6 +32,7 @@
     "dist/",
     "docs/_layouts/",
     "extensions/diffs/assets/viewer-runtime.js",
+    "extensions/browser/chrome-extension/modules/copilot-runtime.js",
     "**/*.json",
     "node_modules/",
     "patches/",
